Added the first few test cases as well as the script to run them.
The tests are in "tests" (conveniently enough) and are placed one per line. Each line consists of a test case, a tab, and then the answer (if you're adding test cases, make sure you do an actual tab character and not a bunch of spaces). test.py (probably should rename it to avoid conflicts) will run each test case and check the answer for accuracy to 10 decimal places. It must be run in Python 2.6 or 3.x.